fix:닉네임 등록과정 - 닉네임 등록 시 반환 값이 사라져서 수정 - 토큰이 키체인에 없을 때를 위한 에러 추가
